supago/requirements.txt
asgiref==3.3.1
boto3==1.16.63
botocore==1.19.63
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
dataclasses==0.8
Django==3.1.5
django-cors-headers==3.7.0
django-storages==1.11.1
djangorestframework==3.12.2
google-api-core==1.26.0
google-auth==1.26.1
google-cloud-vision==2.2.0
googleapis-common-protos==1.52.0
grpcio==1.35.0
idna==2.10
jmespath==0.10.0
libcst==0.3.17
mypy-extensions==0.4.3
mysqlclient==2.0.3
packaging==20.9
Pillow==8.1.0
pkg-resources==0.0.0
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.3
uWSGI==2.0.19.1
